Add on_user_signal to shutdown
