+++ client/src/main.rs
@@ -724,48 +724,6 @@ async fn run_fat(
 	Ok(())
 }
 
-/// This utility function returns a [`Future`] that completes upon
-/// receiving each of the default termination signals.
-///
-/// On Unix-based systems, these signals are Ctrl-C (SIGINT) or SIGTERM,
-/// and on Windows, they are Ctrl-C, Ctrl-Close, Ctrl-Shutdown.
-async fn user_signal() {
-	let ctrl_c = tokio::signal::ctrl_c();
-	#[cfg(all(unix, not(windows)))]
-	{
-		let sig = async {
-			let mut os_sig =
-				tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())?;
-			os_sig.recv().await;
-			std::io::Result::Ok(())
-		};
-
-		tokio::select! {
-			_ = ctrl_c => {},
-			_ = sig => {}
-		}
-	}
-
-	#[cfg(all(not(unix), windows))]
-	{
-		let ctrl_close = async {
-			let mut sig = tokio::signal::windows::ctrl_close()?;
-			sig.recv().await;
-			std::io::Result::Ok(())
-		};
-		let ctrl_shutdown = async {
-			let mut sig = tokio::signal::windows::ctrl_shutdown()?;
-			sig.recv().await;
-			std::io::Result::Ok(())
-		};
-		tokio::select! {
-			_ = ctrl_c => {},
-			_ = ctrl_close => {},
-			_ = ctrl_shutdown => {},
-		}
-	}
-}
-
 mod cli;
 
 pub fn load_runtime_config(opts: &CliOpts) -> Result<RuntimeConfig> {
@@ -880,7 +838,7 @@ pub async fn main() -> Result<()> {
 	};
 
 	// spawn a task to watch for ctrl-c signals from user to trigger the shutdown
-	spawn_in_span(shutdown.with_trigger("user signaled shutdown".to_string(), user_signal()));
+	spawn_in_span(shutdown.on_user_signal("User signaled shutdown".to_string()));
 
 	#[cfg(feature = "crawl")]
 	if let Err(error) = run_crawl(
